+++ packages/web/src/features/search/listReposApi.ts
@@ -4,12 +4,9 @@ import { ListRepositoriesResponse } from "./types";
 import { zoektFetch } from "./zoektClient";
 import { zoektListRepositoriesResponseSchema } from "./zoektSchema";
 import { sew, withAuth, withOrgMembership } from "@/actions";
-import { getAuditService } from "@/ee/features/audit/factory";
-
-const auditService = getAuditService();
 
 export const listRepositories = async (domain: string, apiKey: string | undefined = undefined): Promise<ListRepositoriesResponse | ServiceError> => sew(() =>
-    withAuth((userId, apiKeyHash) =>
+    withAuth((userId, _apiKeyHash) =>
         withOrgMembership(userId, domain, async ({ org }) => {
             const body = JSON.stringify({
                 opts: {
@@ -46,22 +43,6 @@ export const listRepositories = async (domain: string, apiKey: string | undefine
             } satisfies ListRepositoriesResponse));
 
             const result = parser.parse(listBody);
-
-            await auditService.createAudit({
-                action: "query.list_repositories",
-                actor: {
-                    id: apiKeyHash ?? userId,
-                    type: apiKeyHash ? "api_key" : "user"
-                },
-                target: {
-                    id: org.id.toString(),
-                    type: "org"
-                },
-                orgId: org.id,
-                metadata: {
-                    message: result.repos.map((repo) => repo.name).join(", ")
-                }
-            });
 
             return result;
         }, /* minRequiredRole = */ OrgRole.GUEST), /* allowSingleTenantUnauthedAccess = */ true, apiKey ? { apiKey, domain } : undefined)